Move around bitcoin test fns
