+++ crates/l1tx/src/deposit/deposit_tx.rs
@@ -85,13 +85,11 @@ fn parse_deposit_script<'a>(
 mod tests {
 
     use bitcoin::Amount;
+    use strata_test_utils::bitcoin::{build_test_deposit_script, create_test_deposit_tx};
 
     use crate::deposit::{
         deposit_tx::extract_deposit_info,
-        test_utils::{
-            build_test_deposit_script, create_test_deposit_tx, get_deposit_tx_config,
-            test_taproot_addr,
-        },
+        test_utils::{get_deposit_tx_config, test_taproot_addr},
     };
 
     #[test]
